feat!: remove newSwap; replace with constantProduct AMM where needed
