--- packages/zoe/src/contracts/newSwap/types.js
@@ -1,20 +1,0 @@
-/**
- * @typedef {Object} SwapPriceQuote3
- *
- * @property {Amount} amountOut
- * @property {Amount} amountIn
- * @property {Amount} protocolFee
- */
-
-/**
- * @typedef {Object} SwapPriceQuote4
- *
- * @property {Amount} amountOut
- * @property {Amount} amountIn
- * @property {Amount} centralAmount
- * @property {Amount} protocolFee
- */
-
-/**
- * @typedef {SwapPriceQuote3|SwapPriceQuote4} SwapPriceQuote
- */
